+++ neuromapp/CMakeLists.txt
@@ -1,5 +1,4 @@
 add_subdirectory (utils)
 add_subdirectory (hello)
-add_subdirectory (queueing)
 add_subdirectory (coreneuron_1.0)
 add_subdirectory (app)
