+++ neuromapp/app/CMakeLists.txt
@@ -4,7 +4,7 @@ include_directories(${PROJECT_SOURCE_DIR})
 add_executable (app driver.cpp main.cpp)
 target_link_libraries (app
                        hello
-		       queueing
+				       coreneuron10_queueing
                        coreneuron10_kernel
                        coreneuron10_solver
                        coreneuron10_cstep
