+++ neuromapp/coreneuron_1.0/common/data/path.h.in
