--- neuromapp/queueing/CMakeLists.txt
@@ -1,8 +1,0 @@
-add_library (queueing
-	     queue.cpp
-	     main.cpp)
-
-
-install (TARGETS queueing DESTINATION lib)
-install (FILES queue.h pool.h queueing.h spinlock_apple.h spinlock_queue.h thread.h DESTINATION include)
-target_link_libraries(queueing storage coreneuron10_common coreneuron10_kernel coreneuron10_solver)
