+++ test/CMakeLists.txt
@@ -4,6 +4,5 @@ include_directories(${PROJECT_BINARY_DIR})
 include_directories(${PROJECT_SOURCE_DIR})
 add_subdirectory (coreneuron_1.0)
 add_subdirectory (hello)
-add_subdirectory (queueing)
 add_subdirectory (utils)
 add_subdirectory (app)
