+++ test/coreneuron_1.0/kernel.cpp
@@ -32,8 +32,8 @@
 #include <boost/filesystem.hpp>
 
 #include "coreneuron_1.0/kernel/kernel.h" // signature kernel application
-#include "test/coreneuron_1.0/path.h" // this file is generated automatically
-#include "test/coreneuron_1.0/helper.h" // common functionalities
+#include "neuromapp/coreneuron_1.0/common/data/path.h" // this file is generated automatically
+#include "coreneuron_1.0/common/data/helper.h" // common functionalities
 #include "utils/error.h"
 
 namespace bfs = ::boost::filesystem;
